separate bin and dalsoopack classes add getPackedPolys() method
